Move DecodedServiceEnvelope into its own file
